ENH: Adding an option to enable/disable reports.
In this way it should be easier to bootstrap the build
process.  The challenge is that the process of building
the report expect to find the images that are only
generated after the tests are run.

Therefore, a developer should first configure with
reports OFF. Build the project, run the tests, and
then reconfigure to turn Reports ON, then finally
build again, to have the report generated.